+++ src/MoreActionsPopover.tsx
@@ -8,7 +8,7 @@
 
 import React from 'react';
 
-import Button from './Button';
+import Button from './components/Button';
 import Menu from './components/Menu';
 import MenuDivider from './components/MenuDivider';
 import MenuItem from './components/MenuItem';
