+++ src/components/button.less
@@ -6,7 +6,7 @@
  * @copyright 2019-2020 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
-@import './theme/common.less';
+@import '../theme/common.less';
 
 .@{viewer-prefix}-button {
     background-color: transparent;
